requirements.txt
fastapi
uvicorn
python-dotenv
httpx
fastapi-mail
tortoise-orm
